requirements.txt
Django==4.1
djangorestframework==3.13.1
django-rest-auth==0.9.5
django-rest-authtoken==2.1.4

djangorestframework-simplejwt==5.2.0

psycopg2-binary==2.9.3

drf-spectacular==0.23.1
pytest==7.1.2
pytest-django==4.5.2
drf-yasg==1.21.3

